\name{process}
\alias{process}
\title{Process DamID microarray data}
\description{
   Pipeline to process and normalize DamID .pair files.
   Generates quality control and standard format files.
}
\usage{
   process(
      files.table,
      out.path = paste('drosonorm', .dtag(), sep="_"),
      sortby = c("name", "filetype", "none"),
      platform = "GPL8471",
      release = c("dm3R5", "dm2R4"),
      mask = TRUE,
      gff = TRUE,
      wig = TRUE,
      dam = TRUE,
      targets = TRUE,
      plotBias = TRUE,
      plotMA = TRUE,
      plotXY = TRUE,
      plotACF = TRUE,
      plotSample = TRUE,
      graph = "ps",
      cex = ifelse(graph == "ps", 1.8, 1)
   )
}
\arguments{
   \item{files.table}{Path to the target file.}
   \item{out.path}{The name of the directory where to save
     output files (will be created if it does not exist).}
   \item{sortby}{How to organize the subdirectories for output.}
   \item{platform}{The identifier of the used microarray.}
   \item{release}{The genome release for mapping purposes.}
   \item{mask}{Whether to mask plasmid bias (see Details).}
   \item{gff}{Whether to output .gff files.}
   \item{wig}{Whether to output .wig files.}
   \item{dam}{Whether to output .dam files.}
   \item{targets}{Whether to identify target regions in .dam
     files.}
   \item{plotBias}{Whether to plot plasmid bias for quality
     control (see Details).}
   \item{plotMA}{Whether to output MA plots.}
   \item{plotXY}{Whether to output scatter plots (only for
     replicated experiments).}
   \item{plotACF}{Whether to output ACF (auto-correlation) plots.}
   \item{plotSample}{Whether to plot sample regions of the
     profiles.}
   \item{graph}{The type of graph to output.}
   \item{cex}{A general magnification factor for plotting.}
}
\details{
   The \code{file.table} file must contain key information to
   process and normalize the .pair files generated by microarray
   scans. The file must have at least 10 columns, and a header
   with column names "name", "exp1", "exp2", "ctl1", "ctl2",
   "exp1spike", "exp2spike", "ctl1spike", "ctl2spike", "intensity".
   The case and the order are not relevant, but the process will
   fail if any of those column is missing.

   The spikes have to comply to the van Steensel lab scheme. This
   means that they are identified by a number ranging from 1 to
   32 and the letter 'a' or 'b'. The identity of the spike oligos
   is defined in the microarray specifications, but this spiking
   scheme is specific of the Drosophila array GPL8471.

   Addition column are allowed. It is even required in case the
   \code{mask} option is set to \code{TRUE} (see below). 

   The package supports one microarray platform in two releases.
   It is possible to update the package with other platforms by
   adding .rda files containing proper object definition in the
   data directory of the package.

   If the \code{mask} option is set to \code{TRUE}, the signal
   of microarray probes matching the open reading frame of the
   protein being mapped will be set to NA. For this, the gene
   being mapped has to be identified by its canonical FlyBase
   FBgn identifier. This identifier can be anywhere on the line,
   as long as it is properly formatted and that there is only one.

   Plasmid bias can also be used for quality control by plotting
   the DamID signal in the region of the mapped gene. High to 
   saturating signal on the coding sequence of that gene can act
   as a fingerprint for the sample's identity. This complements
   the spiking scheme to trace the samples.

   Targets are identified by a 3-state Hidden Markov Model, which
   is carried out by the \code{HMMt} package. This is done on the
   .dam files, where signals from different probes are averaged
   if they map to the same GATC fragment.

   All the generated files are traced, meaning that the text files
   have a vheader and the plots have a mini signature indicating
   the package version and the date of their production. Tracking
   of the text files is taken care of by the \code{vtrackR} package.
}
\author{
   Ulrich Braunschweig, Aleksey Pindyurin and
   Guillaume Filion <guillaume.filion@gmail.com>
}
